Extract locale validator in single file
